+++ modin/core/execution/dask/implementations/pandas_on_dask/io/io.py
@@ -11,20 +11,10 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
-"""Module houses class that implements ``BaseIO`` using Dask as an execution engine."""
-
-import io
-import pandas
-from pandas.io.common import get_handle
+"""The module holds the factory which performs I/O using pandas on Dask."""
 
 from modin.core.io import BaseIO
 from modin.core.storage_formats.pandas.query_compiler import PandasQueryCompiler
-from modin.core.execution.dask.implementations.pandas_on_dask.dataframe import (
-    PandasOnDaskDataframe,
-)
-from modin.core.execution.dask.implementations.pandas_on_dask.partitioning import (
-    PandasOnDaskDataframePartition,
-)
 from modin.core.io import (
     CSVDispatcher,
     FWFDispatcher,
@@ -43,64 +33,13 @@ from modin.core.storage_formats.pandas.parsers import (
     PandasSQLParser,
     PandasExcelParser,
 )
-from modin.core.execution.dask.common import DaskWrapper
-from distributed.client import default_client
-
-import distributed
-
-
-class SignalActor:  # pragma: no cover
-    """
-    Help synchronize across tasks and actors on cluster.
-
-    For details see: https://docs.ray.io/en/latest/advanced.html?highlight=signalactor#multi-node-synchronization-using-an-actor
-
-    Parameters
-    ----------
-    event_count : int
-        Number of events required for synchronization.
-    """
-
-    def __init__(self, event_count: int):
-        self.events = [distributed.Event() for _ in range(event_count)]
-
-    def send(self, event_idx: int):
-        """
-        Indicate that event with `event_idx` has occured.
-
-        Parameters
-        ----------
-        event_idx : int
-        """
-        self.events[event_idx].set()
-
-    async def wait(self, event_idx: int):
-        """
-        Wait until event with `event_idx` has occured.
-
-        Parameters
-        ----------
-        event_idx : int
-        """
-        await self.events[event_idx].wait()
-
-    def is_set(self, event_idx: int) -> bool:
-        """
-        Check that event with `event_idx` had occured or not.
-
-        Parameters
-        ----------
-        event_idx : int
-
-        Returns
-        -------
-        bool
-        """
-        return self.events[event_idx].is_set()
+from modin.core.execution.dask.common import DaskWrapper, SignalActor
+from ..dataframe import PandasOnDaskDataframe
+from ..partitioning import PandasOnDaskDataframePartition
 
 
 class PandasOnDaskIO(BaseIO):
-    """The class implements interface in ``BaseIO`` using Dask as an execution engine."""
+    """Factory providing methods for performing I/O operations using pandas as storage format on Dask as engine."""
 
     frame_cls = PandasOnDaskDataframe
     query_compiler_cls = PandasQueryCompiler
@@ -109,6 +48,7 @@ class PandasOnDaskIO(BaseIO):
         query_compiler_cls=PandasQueryCompiler,
         frame_cls=PandasOnDaskDataframe,
         base_io=BaseIO,
+        signal_actor=SignalActor,
     )
 
     def __make_read(*classes, build_args=build_args):
@@ -120,6 +60,7 @@ class PandasOnDaskIO(BaseIO):
         return type("", (DaskWrapper, *classes), build_args).write
 
     read_csv = __make_read(PandasCSVParser, CSVDispatcher)
+    to_csv = __make_write(CSVDispatcher)
     read_fwf = __make_read(PandasFWFParser, FWFDispatcher)
     read_json = __make_read(PandasJSONParser, JSONDispatcher)
     read_parquet = __make_read(PandasParquetParser, ParquetDispatcher)
@@ -133,138 +74,3 @@ class PandasOnDaskIO(BaseIO):
 
     del __make_read  # to not pollute class namespace
     del __make_write  # to not pollute class namespace
-
-    @staticmethod
-    def _to_csv_check_support(kwargs):
-        """
-        Check if parallel version of ``to_csv`` could be used.
-
-        Parameters
-        ----------
-        kwargs : dict
-            Keyword arguments passed to ``.to_csv()``.
-
-        Returns
-        -------
-        bool
-            Whether parallel version of ``to_csv`` is applicable.
-        """
-        path_or_buf = kwargs["path_or_buf"]
-        compression = kwargs["compression"]
-        if not isinstance(path_or_buf, str):
-            return False
-        # case when the pointer is placed at the beginning of the file.
-        if "r" in kwargs["mode"] and "+" in kwargs["mode"]:
-            return False
-        # encodings with BOM don't support;
-        # instead of one mark in result bytes we will have them by the number of partitions
-        # so we should fallback in pandas for `utf-16`, `utf-32` with all aliases, in instance
-        # (`utf_32_be`, `utf_16_le` and so on)
-        if kwargs["encoding"] is not None:
-            encoding = kwargs["encoding"].lower()
-            if "u" in encoding or "utf" in encoding:
-                if "16" in encoding or "32" in encoding:
-                    return False
-        if compression is None or not compression == "infer":
-            return False
-        if any((path_or_buf.endswith(ext) for ext in [".gz", ".bz2", ".zip", ".xz"])):
-            return False
-        return True
-
-    @classmethod
-    def to_csv(cls, qc, **kwargs):
-        """
-        Write records stored in the `qc` to a CSV file.
-
-        Parameters
-        ----------
-        qc : BaseQueryCompiler
-            The query compiler of the Modin dataframe that we want to run ``to_csv`` on.
-        **kwargs : dict
-            Parameters for ``pandas.to_csv(**kwargs)``.
-        """
-        if not cls._to_csv_check_support(kwargs):
-            return BaseIO.to_csv(qc, **kwargs)
-
-        signals = (
-            default_client()
-            .submit(SignalActor, len(qc._modin_frame._partitions) + 1, actors=True)
-            .result()
-        )
-
-        def func(df, **kw):
-            """
-            Dump a chunk of rows as csv, then save them to target maintaining order.
-
-            Parameters
-            ----------
-            df : pandas.DataFrame
-                A chunk of rows to write to a CSV file.
-            **kw : dict
-                Arguments to pass to ``pandas.to_csv(**kw)`` plus an extra argument
-                `partition_idx` serving as chunk index to maintain rows order.
-            """
-            partition_idx = kw["partition_idx"]
-            # the copy is made to not implicitly change the input parameters;
-            # to write to an intermediate buffer, we need to change `path_or_buf` in kwargs
-            csv_kwargs = kwargs.copy()
-            if partition_idx != 0:
-                # we need to create a new file only for first recording
-                # all the rest should be recorded in appending mode
-                if "w" in csv_kwargs["mode"]:
-                    csv_kwargs["mode"] = csv_kwargs["mode"].replace("w", "a")
-                # It is enough to write the header for the first partition
-                csv_kwargs["header"] = False
-
-            # for parallelization purposes, each partition is written to an intermediate buffer
-            path_or_buf = csv_kwargs["path_or_buf"]
-            is_binary = "b" in csv_kwargs["mode"]
-            csv_kwargs["path_or_buf"] = io.BytesIO() if is_binary else io.StringIO()
-            storage_options = csv_kwargs.pop("storage_options", None)
-            df.to_csv(**csv_kwargs)
-            csv_kwargs.update({"storage_options": storage_options})
-            content = csv_kwargs["path_or_buf"].getvalue()
-            csv_kwargs["path_or_buf"].close()
-
-            # each process waits for its turn to write to a file
-            # DaskWrapper.materialize(signals.wait(partition_idx))
-            signals.wait(partition_idx).result()
-
-            # preparing to write data from the buffer to a file
-            with get_handle(
-                path_or_buf,
-                # in case when using URL in implicit text mode
-                # pandas try to open `path_or_buf` in binary mode
-                csv_kwargs["mode"] if is_binary else csv_kwargs["mode"] + "t",
-                encoding=kwargs["encoding"],
-                errors=kwargs["errors"],
-                compression=kwargs["compression"],
-                storage_options=kwargs.get("storage_options", None),
-                is_text=not is_binary,
-            ) as handles:
-                handles.handle.write(content)
-
-            # signal that the next process can start writing to the file
-            # DaskWrapper.materialize(signals.send(partition_idx + 1))
-            signals.send(partition_idx + 1).result()
-            # used for synchronization purposes
-            return pandas.DataFrame()
-
-        # signaling that the partition with id==0 can be written to the file
-        # DaskWrapper.materialize(signals.send(0))
-        signals.send(0).result()
-        # Ensure that the metadata is syncrhonized
-        qc._modin_frame._propagate_index_objs(axis=None)
-        result = qc._modin_frame._partition_mgr_cls.map_axis_partitions(
-            axis=1,
-            partitions=qc._modin_frame._partitions,
-            map_func=func,
-            keep_partitioning=True,
-            lengths=None,
-            enumerate_partitions=True,
-            max_retries=0,
-        )
-        # pending completion
-        DaskWrapper.materialize(
-            [part.list_of_blocks[0] for row in result for part in row]
-        )
